Don't schedule PowerPC jobs by default, for now
We have no PowerPC workers on prod, but now we merged the PPC
changes, PPC jobs are getting scheduled and piling up. We don't
want that. For now, let's remove ppc64 images from the stock
WANTED list, and to keep PowerPC tests running on staging, I'll
give it an override file for the WANTED list via ansible.