Use new _ONLYOBSOLETESAME behaviour for composes and updates
I just sent a PR upstream that adds a new _ONLYOBSOLETESAME
scheduling param. This is somewhere between default behaviour
and _NOOBSOLETEBUILD; pending jobs may be canceled when a new
'ISO' POST occurs, but only jobs for the exact same BUILD, not
jobs with the same DISTRI, VERSION, FLAVOR and ARCH but a
different BUILD (as is the default behaviour). This means we
can still re-POST an image to cancel existing jobs for that
image and re-start them, but jobs for a modular compose won't
be canceled when a non-modular compose of the same release
appears, and jobs for one update won't be canceled when a new
update for the same release appears.