Discover and pass in Rawhide release number as `RAWREL` var
There is a release number associated with Rawhide in some ways:
notably, in this case, for RPM GPG keys. We want to use this
release number as the `--releasever` param for upgrade tests
(rather than the string 'rawhide') to avoid #1531356, so we
need to discover it and pass it in for the tests to use: let's
do that. If Branched exists, the Rawhide release number is one
higher than that number, otherwise it's one higher than the
current stable release: this use of fedfind's
`get_current_release` should do the trick.

Signed-off-by: Adam Williamson <awilliam@redhat.com>