Adjust update test scheduling for FreeIPA upgrade tests
This commit includes a few changes to scheduling such that we
can run tests involving	a system upgrade on updates. It	allows
scheduling of the new 'server-upgrade' flavor, adjusts the
whitelist for that, avoids scheduling upgrade tests for	updates
to the oldest current stable release (as we don't support
upgrades from EOL releases and we don't	have the disk images
available), passes CURRREL in for these tests (which is used
to decide which base disk image to boot from and thus which
release to upgrade *from*), and updates the tests for all these
changes.

Signed-off-by: Adam Williamson <awilliam@redhat.com>