Fix duplicate detection for composes with identical ISO names
The scheduling duplicate detection assumes that ISO names will
always be unique. This is no longer the case, however, as now
we have identical filenames in two-week Atomic nightly, updates
and updates-testing composes: for these we wind up only running
tests on whichever compose happens first, and the jobs for the
others are incorrectly skipped by the duplicate detection. So,
add a BUILD check to the duplicate detection to avoid this.

Signed-off-by: Adam Williamson <awilliam@redhat.com>