Fix test_jobs_from_compose_unsupported
In recent fedfind, updates and updates-testing composes are now
actually supported, as they contain images and we want to test
them. There are no longer any 'real' composes that fedfind would
treat as unsupported, currently...but we can fake one up easily
enough, so let's do that, just to test the mechanism.

Signed-off-by: Adam Williamson <awilliam@redhat.com>