Don't pass HDD_1 to updates-installer flavor
We only want it for the installer creation test itself, any
subsequent install tests do *not* want it. As it's impossible
to override a set HDD_1 to an unset HDD_1 (openQA doesn't handle
`HDD_1=""` correctly), let's not set it here then override it
for the install tests - let's just not set it here, and have
the installer build test set it in `templates-updates`.

Signed-off-by: Adam Williamson <awilliam@redhat.com>