Allow filtering by flavor when scheduling a compose
You can already do this by using a WANTED file, but that's a lot
of work if you just quickly want to schedule jobs only for one
ISO or whatever. It's much easier to just be able to pass in a
comma-separated list of flavors on the CLI, like this.

Signed-off-by: Adam Williamson <awilliam@redhat.com>