Handle scheduling update live image create and install tests
This also rethinks the flavor naming strategy a bit, to use
names that are more specific to the image being built and
more similar to the flavors used for compose tests. I tried
using the *actual same* flavors the compose tests use, but
that leads to jobs being scheduled for updates as if they
were composes; there's a feature you can use to prevent that,
but it just seemed like it was getting a bit complicated and
I preferred this approach in the end, for now.

Signed-off-by: Adam Williamson <awilliam@redhat.com>