Rename updates-testing images on download to avoid name clash
I noticed yesterday that the filenames of images in updates and
updates-testing composes are identical. For instance, both the
updates and updates-testing composes for Fedora 29 for 20190220
contain a 'Fedora-AtomicHost-ostree-x86_64-29-20190220.0.iso'.
This is a problem, because openQA assets are all stored together
in one directory, there's no subdirectory per BUILD or anything
like that. So all the time we've been testing these composes,
we've actually been inadvertently just testing the same images
twice - each day we'll have only actually tested one of the
composes and not tested the other one at all.

To try and deal with this, we'll rename the image files for the
updates-testing composes on download. The openQA asset download
code has a mechanism for doing this: if you pass for e.g. both
ISO_URL and ISO, the file found at the URL will be downloaded
under the name give as 'ISO'. Then, when we report results to
resultsdb, we'll reverse that munging, so the 'item' name we
report is the *original* file name, not the altered one.

Signed-off-by: Adam Williamson <awilliam@redhat.com>