Convert from fedmsg to fedora-messaging
This converts all the consumers (for job scheduling and result
reporting) from fedmsg to fedora-messaging. There are some
differences in configuration: things that were previously set
in `schedule.conf` are now set in the consumer config file. You
will need to follow the instructions in the README to set up
the fedora-messaging consumers.

Signed-off-by: Adam Williamson <awilliam@redhat.com>