Pass list of NVRs as variable when scheduling updates (#78)
This is necessary for us to be able to publish spec compliant
CI Messages, as the spec requires the messages to include the
list of NVRs in the update as tested. This will be accompanied
by a change to os-autoinst-distri-fedora to actually download
these NVRs, not just whatever is in the update at the time the
test runs.

Signed-off-by: Adam Williamson <awilliam@redhat.com>