Avoid scheduling update tests for Rawhide updates
We don't seem to get messages published for *most* Rawhide
updates (those have started existing recently for CI purposes),
but we do get messages for a *few*, and I caught one case where
we tried to run update tests and it all went pear-shaped:

https://openqa.fedoraproject.org/tests/overview?build=Update-FEDORA-2019-b63884f469&version=32&distri=fedora&groupid=2

This should hopefully protect us against that. Note this should
probably be used with fedfind 4.2.8 or later or else the stale
collections cache will cause us to not schedule updates for a
newly-Branched release.

Signed-off-by: Adam Williamson <awilliam@redhat.com>