Hack 31 to be considered 'stable' for update scheduling
This is an awful hack but I need to fix it quick. This whole
mechanism doesn't cover the time near the release date where we
send out the 'final' versioned fedora-release package, as at
that point, the live image built by live_build will be a 'final'
one due to the fedora-release package version, but the tests
still expect it to be a 'pre-release' image. I can't think of a
quick correct fix for this, so for now let's just hack 31 into
the list of 'stable' releases used when deciding whether to set
the DEVELOPMENT var when scheduling update tests.

Signed-off-by: Adam Williamson <awilliam@redhat.com>