Tweaks to avoid a spurious error in checkwiki
We only report for QA:Testcase_workstation_core_applications if
the workstation_core_applications module passes. That module is
only scheduled *at all* when the desktop is GNOME; when running
the test on KDE we just don't schedule it. This tweaks the passed
test case calculation in report.py and checkwiki itself to be
aware of this, so we don't get a logged warning or a spurious
'No ResultRow found' result from checkwiki for this scenario.

Signed-off-by: Adam Williamson <awilliam@redhat.com>