consumer: allow configuration of arches for update tests
This adds a config setting for the job scheduling consumer which
allows configuration of the arches for which update jobs will be
scheduled. This is to allow us to enable testing of updates on
ppc64le in Fedora staging as a test; if that goes well, we will
enable them in production, and also later enable update tests on
aarch64 once we have more workers.

The new setting is required, so if anyone else is actually using
these consumers, you'll need to update your configuration.

Signed-off-by: Adam Williamson <awilliam@redhat.com>