Move module under src/, add pyproject.toml, drop pytest-cov
This moves the Python module under src/ , for reasons discussed
at https://hynek.me/articles/testing-packaging/ : essentially
it's good practice to *not* have the module directly importable
from the top-level working directory.

The convenience wrapper for running the CLI direct from the
top-level directory and the `checkwiki` test script are updated
to still work with the module moved like this. We also update the
tox (CI) configuration to work with it, by building an sdist and
testing that, rather than testing the contents of the working
directory: this is a stronger test (as we ultimately deploy via
sdist, so we need to make sure the sdist we build works). We also
update setup.py for the change.

We add pyproject.toml to express that setuptools-git is needed
to build the sdist correctly, and we switch to running coverage
and having it call pytest versus using pytest-cov as this is
necessary for coverage diffing to work correctly with the layout
change.

Signed-off-by: Adam Williamson <awilliam@redhat.com>