Update awful 'stable' hack to F32
Well, I've done this for two releases now, so I guess it's a
tradition! :(

Same problem as before: after fedora-release gets bumped to -1
or higher, the live image that we build in the updates tests no
longer has the pre-release note, but we are still expecting it
in the install test because the release is not "stable" yet so
far as our source of truth (now Bodhi, not collections) knows.
Even with a whole cycle to think about it, I haven't come up
with a better way to fix this, unfortunately.

Signed-off-by: Adam Williamson <awilliam@redhat.com>