Redo release vars: add UP{1,2}REL and LABEL, don't set DEVELOPMENT
The DEVELOPMENT var is pretty problematic and sometimes a lie.
I'm rejigging the tests to not use it, and improve their logic.
But this will mean we always want CURRREL to be set and always
be the actual current release (we don't want it to be a lie for
update tests sometimes). So we drop DEVELOPMENT, invent UP1REL
and UP2REL to be specifically for the purpose of defining the
upgrade source releases, and always set CURRREL, PREVREL, UP1REL
and UP2REL for all tests including update tests. We also change
the logic for Rawhide upgrade test source versions: we had just
been using CURRREL and PREVREL, which meant that when there is
a Branched, we're not really testing the right thing, we're
testing upgrades from Rawhide-2 and Rawhide-3. This is now
changed so we always test Rawhide-1 and Rawhide-2, even if
Rawhide-1 is Branched. We also add a LABEL var which is set if
the compose has a label, this is also to help the tests decide
whether to expect prerelease stuff.

Signed-off-by: Adam Williamson <awilliam@redhat.com>