Fix flow errors in resultsdb reporting
This damn 'retry on error' loop sure has caused more problems
than it ought. Latest one: it short-circuits if you try to report
more than one result at a time, because of that `return`. We have
to get that out of there. This should do the trick, while still
raising the last error if we can.

Signed-off-by: Adam Williamson <awilliam@redhat.com>