Add job scheduling and ResultsDB reporting for Fedora CoreOS
This adds the ability to schedule jobs and report results to
ResultsDB for Fedora CoreOS builds. These are not Pungi composes
and so don't have productmd-style metadata or messages, so we
handle them entirely separately. Currently you can only schedule
jobs for the current build from one of the streams, and only for
x86_64 and the "live" ISO. This requires a new version of
resultsdb_conventions that adds conventions for Fedora CoreOS
build results.

Also refactors test_cli a bit as it was confusing that some tests
were in classes and others weren't. Think I did this to make
pylint happy but I'd rather make humans happy than a script,
really...

Signed-off-by: Adam Williamson <awilliam@redhat.com>