Drop 'SUBVARIANT_OR_ARM' replacement, enable aarch64 in checkwiki
I'm planning to enable aarch64 testing in openQA prod. This is
part of the groundwork. We 'enable' aarch64 in checkwiki, i.e.
check wiki result submission concordance with the test templates
and the wiki matrix templates for aarch64 as well as x86_64. I
made some changes to the matrix templates themselves at the same
time which mean checkwiki now almost passes, and I mean to make
some changes to the test templates that will fix the remaining
issue. We make one small change to the test suite config here to
match a change made in the wiki.

Signed-off-by: Adam Williamson <awilliam@redhat.com>