Schedule disk image assets as HDD_2
For the Cloud disk image, we schedule it with the image URL as
HDD_1_URL and run the Base tests on it. But for all other flavors,
those tests run on a disk image produced by install_default_upload
test, so the test suites specify HDD_1. Problem with that is,
the HDD_1 value from the test suite gets used as the file name
for the downloaded asset (instead of it being generated from
the HDD_1_URL value like it would be if HDD_1 were not set),
and because that name doesn't change between builds, we're not
actually redownloading the file from the new compose as we should
be; we downloaded it just once, back in November, and every day
we re-test that same file, because openQA looks for the same file
as it did the day before, and that file is still there, so it
just goes ahead and uses it...

Fixing this with overrides in the templates is somewhere between
difficult and impossible because if you just override HDD_1 to an
empty string, it doesn't cause the autonaming behaviour to be
used, it instead is taken as a signal you don't want to attach
the asset at all. This is because in *another* case, *that* is
the behaviour we want, so we can't change that. We would have to
do something pretty ugly to make it so we can kinda use overrides
to produce two different behaviours in the two cases where we want
different things to happen.

So let's dodge the issue this way: when downloading disk image
assets, let's just set them as HDD_2 not HDD_1 (as we already do
for ARM disk images, in fact). We will have the product in the
templates set +HDD_1="%HDD_2%" to 'clone' the HDD_2 value (after
it has been parsed out of HDD_2_URL) to HDD_1.

Signed-off-by: Adam Williamson <awilliam@redhat.com>