Switch FCOS scheduling to use build metadata, add consumer (#85)
There are now messages (coreos.build.state.change) we can use
to trigger tests of FCOS builds. This includes builds that aren't
published to the public streams, so in order to test those, we
change the triggering code to run from the 'build URL' (which
is provided in these messages), using a per-build metadata file
(`meta.json`) which should always be present. To manually trigger
tests you need to know this build URL. This change also allows
for scheduling for aarch64, though the templates aren't yet set
up for this.

Signed-off-by: Adam Williamson <awilliam@redhat.com>