Add scheduling for new Workstation upgrade test flavors
In order to run post-install tests on upgraded systems, we're
moving the Workstation upgrade test (only that, for now) into
its own flavor:

https://pagure.io/fedora-qa/os-autoinst-distri-fedora/issue/226
https://pagure.io/fedora-qa/os-autoinst-distri-fedora/pull-request/253

this handles the scheduling for those flavors. We add them into
the list we get from parsing the images in the compose, before
the filtering happens so they can be filtered out if desired.
This also requires the tests to be updated, and an improvement
to the logic for raising an exception if we don't find images.

Signed-off-by: Adam Williamson <awilliam@redhat.com>