Set up wiki reporting for eog, evince and gedit tests
These can report QA:Testcase_desktop_app_basic for the respective
app 'types'. Since the test name is the app name, but the matrix
has app 'types' like "document viewer", we need a translation
layer via the magic value replacement stuff. We also need to
include desktop_terminal in this, as we'd previously been just
assuming that was the *only* desktop_app_basic report we'd make.

Signed-off-by: Adam Williamson <awilliam@redhat.com>