Don't schedule upgrade flavors for non-'full' composes
After deploying the previous commit to schedule the new special
upgrade flavors, a problem became apparent - we're scheduling
them even for things like the respin composes and nightly stable
composes, where they don't make sense.

I thought of a few different filters, but I think I like this
one: we rely on fedfind's https_url_generic property as an
indicator that the compose has a 'generic' tree. If a compose
has a 'generic' tree it's probably possible/sensible to test
upgrades to it. We could've just gone with a dumber filter that
only scheduled these flavors on non-Respin composes for the
'Fedora' dist, and if this still turns out to be problematic,
I'll fall back to that. Note this will only work properly with
fedfind 5.0.0 or higher, with older fedfind we'll still schedule
for nightly stable composes.

Signed-off-by: Adam Williamson <awilliam@redhat.com>