Handle job not having DESKTOP setting in _get_passed_tcnames
This ought to be a safe assumption - any test with the
"workstation_core_applications" module ought to be a desktop
test at least - but on lab it isn't; job 1288959 is from a time
when lruzicka was experimenting with upgrade test scheduling
and it didn't get a DESKTOP setting. So we need to handle this
unexpected case.

Signed-off-by: Adam Williamson <awilliam@redhat.com>