Schedule update tests for Rawhide
This requires various tweaks, and also requires handling non-
retrigger bodhi.update.status.testing messages, as this is all
we get for Rawhide updates (there is no request.testing for
them). Currently we have to retrieve the update dict from
Bodhi to find out whether the update is critpath when doing
this, as the message doesn't tell us.

Signed-off-by: Adam Williamson <awilliam@redhat.com>