Report base_initial_setup on Workstation and Silverblue (#100)
This requires a dumb special hack that I'm sad about, but for
this one case, I can't justify adding a generic mechanism. If
we find anything else like this, I will (something like letting
the test case name be a tuple or something dumb like that).

install_default_upload for Silverblue and Workstation does not
create a user or set a root password, as the installers are set
to suppress those spokes. So they're effectively the same as
install_no_user, which is why we don't run install_no_user
separately for those flavors. But that means we don't report
initial_base_setup for those flavors, because that reporting is
tied to install_no_user on the assumption *that* is the test
that tests installing without creating a user. To solve that,
use this ugly hack which just sticks the test case on the end
of the list for those two subvariants.

Signed-off-by: Adam Williamson <awilliam@redhat.com>