Use form data, not query params, for the ISO POST for updates
As noted at https://progress.opensuse.org/issues/119200 , we
ran into a fun problem with a KDE update with 200+ packages in
it: the request URI (with all the parameters encoded in it) was
longer than the ~8k limit that's the default for Apache and
also, apparently, something deep in Perl that caused openQA to
barf on the request even when I changed Apache's config.

It turns out, all the bits are already in place in openQA itself
and openQA-python-client to let us send the parameters as form
data rather than encoded in the URI. All we have to do is pass
the dict as the `data` arg rather than the `params` arg. So,
let's do that! This should hopefully mean we never have this
problem again even if someone makes an even bigger update.

Not doing the same (at least for the moment) for other queries
as I don't think there's any chance of them producing URIs this
long.

This also drops some stray debug `print()` statements in the
tests which I noticed while updating them.

Signed-off-by: Adam Williamson <awilliam@redhat.com>