Detect release for update tests if not explicitly provided
It's kind of a pain that you have to provide the release number
for update tests when scheduling them manually. Initially I did
it this way so the scheduler didn't have to get the update
metadata from Bodhi, but it's been doing that for every update
for years now to provide the ADVISORY_NVRS list. Since we're
doing that anyway, we may as well also pull the release number
from the same metadata unless it's explicitly provided. You
still have to provide the release number when scheduling tests
for a Koji task; we could probably get the info from Koji, but
it's more painful than Bodhi (hello, XMLRPC!) and we don't do it
already so I'm not wasting time adding it.

It doesn't inherently make much sense to still allow the release
number to be passed manually, but I'm leaving it around so we
don't have to change the scheduler or tests.

Signed-off-by: Adam Williamson <awilliam@redhat.com>