Only schedule necessary tests for grouped critpath updates
This builds on a bunch of enhancements to bodhi and greenwave
I've been referring to as "grouped critical path". As of today,
newly-created and edited updates aren't just "critical path" or
"not critical path", but include info on which critical path
groups they contain packages from. The greenwave policies have
been updated to only require relevant tests for the groups the
update is a part of. So we can refine scheduling to only schedule
the necessary tests too.

This is only done when flavors aren't explicitly specified. If
the update is an older one that doesn't list the critpath groups
but is just tagged as being critical path, we'll still schedule
all flavors.

Signed-off-by: Adam Williamson <awilliam@redhat.com>