Support reporting resultsdb "results" for non-completed jobs
We decided a while back to allow reporting non-completed "results"
to resultsdb as a mechanism for tracking automated test execution
state: https://pagure.io/fedora-infrastructure/issue/8989
resultsdb explicitly supports the QUEUED and RUNNING "outcomes",
which indicate a test that is scheduled or currently running. CI
already reports such results, but openQA does not yet. Adding
this will allow us to adapt Bodhi to show scheduled/running
tests for updates, as well as completed ones, which is nicer
than just saying the test results are missing.

Some more work is needed on openQA side for this to work fully -
there is currently no 'job started running' event, and openQA
does not emit the 'job_duplicate' event when a job is restarted
via the RETRY auto-restart mechanism, which we use for update
tests. But this will at least give us QUEUED results for the
initial creation of jobs, and manual restarts.

Signed-off-by: Adam Williamson <awilliam@redhat.com>