Update test compose data and adjust schedule tests
The test 'real' compose data we use in some of the scheduling
tests is *very* old and not much like current composes - it has
32-bit ARM images in it, it *doesn't* have 64-bit ARM images in
it, and it doesn't have Silverblue. So let's rip it and replace
it with data from today's Rawhide compose. This requires a lot
of tweaks to the tests themselves, too, as the set of tests we
schedule from this data is quite different from the set of tests
we schedule from the old data.

Also a few minor layout tweaks to avoid overlong lines.

Signed-off-by: Adam Williamson <awilliam@redhat.com>