Drop complex legacy 'universal' scheduling
As described in more detail in the os-autoinst-distri-fedora
commit that goes with this, the complex 'universal' scheduling
is unnecessary now. We're keeping a rump 'universal' flavor but
all it has in it is upgrade tests; it can be handled exactly
like the "Workstation-upgrade" flavor that already exists, no
image needs to be associated with the flavor. We can drop all
the complex 'score' stuff.

Signed-off-by: Adam Williamson <awilliam@redhat.com>