consumer: don't re-test on update edit if builds didn't change
There's a pending change in Bodhi that will make this easier:
https://github.com/fedora-infra/bodhi/pull/5237
but we can actually do it right now without *too* much effort,
so let's. This should avoid rescheduling update tests when the
update is edited, but the edit didn't change the builds (e.g.
an edit to the description or the associated bugs). We used to
intentionally *not* do this so you could edit the update as a
hacky way to forcibly re-trigger tests (at least, if you were
the update's owner or a provenpackager...), but now we have a
button for re-triggering the tests, we don't need that any more,
and now we're reporting 'queued'/'running' results, it's good to
avoid rerunning tests when not really necessary.

Signed-off-by: Adam Williamson <awilliam@redhat.com>