consumer: report new 'results' for all restarted jobs
When we restart a job, it may result in parent or child jobs also
being restarted, and we want to report "results" (which will be
QUEUED) for all of them, not just the one that was directly
restarted. Otherwise we wind up in a state where the related jobs
are now queued/running in openQA, but still shown as completed
with the pre-restart result in Bodhi.

Signed-off-by: Adam Williamson <awilliam@redhat.com>