Update test scheduling: build and attach a repo ISO
This is a new approach to handling the update and workaround
repos for update tests. Instead of each test run recreating
these for itself, have the scheduler create them just once when
scheduling an update test, and stuff them in an ISO and attach
it to the tests. The tests then just have to mount the ISO and
set up the repository config.

Signed-off-by: Adam Williamson <awilliam@redhat.com>