Don't produce incomplete updates image for large updates
Whoops. We're feeding `nvrs` to `_build_update_image`, but this
loop was overwriting it with just the last chunk of NVRs, when
there are more than 20. That means we get an incomplete update
image for updates with more than 20 packages. Let's not do that!

Serves me right for pushing this out without writing tests...

Signed-off-by: Adam Williamson <awilliam@redhat.com>