Ensure HOME env var is set in _do_package_download
There was a weird problem on the Fedora staging instance (but not
the prod instance): when run via a fedora-messaging consumer -
as is the usual case - we were crashing here because
os.environ["HOME"] wasn't set. The Bodhi client bindings assume
it will be, so the bodhi CLI crashes if it's not.

When run directly from the CLI it worked fine. I have no idea
how it's somehow not set when running via a consumer, but we need
it to work, so we have to work around it.

Note: os.path.expanduser tries to use os.environ["HOME"] in the
first instance, but if isn't set, it works via pwd lookup, as
described in the upstream docs:
https://docs.python.org/3/library/os.path.html#os.path.expanduser

Signed-off-by: Adam Williamson <awilliam@redhat.com>