Tweak handling of ODCS composes with None compose ID, add tests
This tweaks the emergency fix tflink added for ODCS composes
with the compose ID specifically set to None (thanks!) When the
ID is None we should skip scheduling, we shouldn't continue -
we *only* want to schedule when the compose ID is set and it's
set to Fedora-ELN-something. This makes the logic more compact
and adds a test to make sure we do what we intend.

Signed-off-by: Adam Williamson <awilliam@redhat.com>