Report results for chained children of failed tests (#107)
I'm a bit concerned this might have unforeseen consequences, but
it's an attempt to solve the problem that when a parent test
fails, we don't report any result for chained children (which
get cancelled by openQA without any event being emitted). This
means Bodhi thinks they're queued forever (or until we manually
restart the tests and get them to pass somehow), as the last
result it 'sees' for the test is the QUEUED one.

This was less of an issue before we implemented reporting of
QUEUED results because the test would just show up as 'missing'.

Signed-off-by: Adam Williamson <awilliam@redhat.com>