Fix scheduling UPDATETL flavors when update is critpath
When an update is in the critical path, we don't schedule any
flavors from UPDATETL for the update even if we should. This was
broken by the critical path group support a while back, but I
only just noticed (podman updates aren't running the 'consumer'
tests).

This fixes that, with a bit of refactoring because fixing it in
the current design was rather ugly and awkward. This makes things
rather cleaner.

It also fixes the CLI path to work the same way here - previously
we never added UPDATETL groups when scheduling critpath updates
from the CLI, but there's no reason we shouldn't - and makes
scheduling slightly more efficient (we don't need to grab the
update dict from Bodhi on the consumer path, the consumer just
passes it in from the message).

Signed-off-by: Adam Williamson <awilliam@redhat.com>