Try not to report failures for tests that will be retried (#105)
It seems like at present, when an update test fails and is
automatically retried (by openQA "RETRY" logic), we report the
initial failure. We don't report the retry starting, though,
because we don't get an openQA event when that happens:
https://progress.opensuse.org/issues/123625
the consequences of this aren't great. The update gating status
goes to 'failed', then only belatedly flips back to 'waiting'
next time it gets updated for any other reason. The Automated
Tests table in Bodhi shows the test as failed until the retry
completes.

We already have logic to try and avoid reporting failures when
the job has a clone, but that's not helping here. My theory is
that the cloning happens just *after* we report the failure. So,
when a job looks like one that should get retried (it has RETRY
set and is not a clone itself), let's try waiting a few seconds
and re-retrieving the job dict before proceeding. I hope that
after the wait, the job dict will have the 'clone_id' entry and
so we'll then not report the result.

Signed-off-by: Adam Williamson <awilliam@redhat.com>