Simplify update test scheduling for Bodhi 8 changes
Bodhi merged a change I wrote to message publication:
https://github.com/fedora-infra/bodhi/pull/5538
it now publishes a single message (UpdateReadyForTesting) in all
the situations when a test system would likely want to test or
re-test the update (update created, update edited with changed
builds, or a re-trigger request sent). This makes it much easier
to handle scheduling, so let's do that.

We now just trigger tests (with force=True) any time we see that
message. This changes how re-trigger requests are handled
slightly, but I couldn't really justify keeping that special case
any more. It was originally necessary because these messages
didn't have an update dict, then they grew one but I kept it
because it felt kinda neat, but now it just sticks out and it's
inconsistent with what we do when updates are edited anyway, so
let's just junk it and always do the same.

Signed-off-by: Adam Williamson <awilliam@redhat.com>