Set TOOLBOX_IMAGE var when we find a toolbox image in the compose
Part of addressing:
https://pagure.io/fedora-qa/issue/766

we want to test the toolbox images that are produced as part of
each compose. But we don't test these like we test other images,
by making them openQA test assets and having them attached to
the VM as virtual disks - we just want to pass the URLs through
to the toolbox test so it can download them and use them. This
should set the TOOLBOX_IMAGE var when scheduling for an arch if
we could find a toolbox image of that arch in the compose being
scheduled. There's a corresponding change to the tests which
does the right stuff when this variable is set.

I had to fake the test images.json data up a bit to test this,
as the compose we took the file from is old enough that it did
not have these images. Instead of updating all the test data to
a newer compose and changing a bunch of things in the tests, I
just copied in the dicts from a newer compose and adjusted the
dates...

Signed-off-by: Adam Williamson <awilliam@redhat.com>