Handle attempt to mark candidate as 'important' failing
This can fail if there is no group 1, which is possible on an
instance where the admin has fiddled with the groups a bit. So,
let's let it fail.

Really we should do something through the API to find out which
group is called 'fedora' and use that, but that's a bit harder,
let's just do this for now.

Signed-off-by: Adam Williamson <awilliam@redhat.com>