Add a new CLI subcommand to find and fix blocked updates
There've been several cases recently where the final result of
an openQA test was never reported to resultsdb. I have not yet
been able to figure out why. This is a workaround: it finds
updates which seem to be stuck in gating, and finds any QUEUED
or RUNNING results for them which are more than 8 hours old. It
checks with openQA for the current state of these tests, and
logs to the console. With --report , it will also report results
for tests which now seem to be finished.

There is intentionally no real configuration for this subcommand
because it's only really useful in production. So it's hardcoded
to use the prod URLs for Bodhi, Greenwave and openQA, and it
just reports to the default resultsdb (with the intent that it
should be run on the prod openQA server, which will report to
prod resultsdb).

Signed-off-by: Adam Williamson <awilliam@redhat.com>